+++ java/core/libjoynr/src/main/java/io/joynr/messaging/routing/LibJoynrMessageRouter.java
@@ -79,18 +79,22 @@ public class LibJoynrMessageRouter extends AbstractMessageRouter {
                                  @Named(SystemServicesSettings.LIBJOYNR_MESSAGING_ADDRESS) Address incomingAddress,
                                  @Named(SCHEDULEDTHREADPOOL) ScheduledExecutorService scheduler,
                                  @Named(ConfigurableMessagingSettings.PROPERTY_SEND_MSG_RETRY_INTERVAL_MS) long sendMsgRetryIntervalMs,
+                                 @Named(ConfigurableMessagingSettings.PROPERTY_MESSAGING_MAXIMUM_PARALLEL_SENDS) int maxParallelSends,
                                  MessagingStubFactory messagingStubFactory,
                                  MessagingSkeletonFactory messagingSkeletonFactory,
                                  AddressManager addressManager,
-                                 MulticastReceiverRegistry multicastReceiverRegistry) {
+                                 MulticastReceiverRegistry multicastReceiverRegistry,
+                                 BoundedDelayQueue<DelayableImmutableMessage> messageQueue) {
         // CHECKSTYLE:ON
         super(routingTable,
               scheduler,
               sendMsgRetryIntervalMs,
+              maxParallelSends,
               messagingStubFactory,
               messagingSkeletonFactory,
               addressManager,
-              multicastReceiverRegistry);
+              multicastReceiverRegistry,
+              messageQueue);
         this.incomingAddress = incomingAddress;
     }