+++ java/core/libjoynr/src/main/java/io/joynr/runtime/AbstractRuntimeModule.java
@@ -92,6 +92,10 @@ import joynr.system.RoutingTypes.Address;
 
 abstract class AbstractRuntimeModule extends AbstractModule {
 
+    // TODO start skeletons and MQTT in a separate thread pool
+    private static final int MQTT_THREADS = 4;
+    private static final int MAX_SKELETON_THREADS = 5;
+
     MapBinder<Class<? extends Address>, AbstractMiddlewareMessagingStubFactory<? extends IMessagingStub, ? extends Address>> messagingStubFactory;
     MapBinder<Class<? extends Address>, IMessagingSkeleton> messagingSkeletonFactory;
     @SuppressWarnings("URF_UNREAD_FIELD")
@@ -171,8 +175,8 @@ abstract class AbstractRuntimeModule extends AbstractModule {
         ThreadFactory schedulerNamedThreadFactory = new ThreadFactoryBuilder().setNameFormat("joynr.MessageScheduler-scheduler-%d")
                                                                               .setDaemon(true)
                                                                               .build();
-        ScheduledThreadPoolExecutor scheduler = new ScheduledThreadPoolExecutor(maximumParallelSends,
-                                                                                schedulerNamedThreadFactory);
+        ScheduledThreadPoolExecutor scheduler = new ScheduledThreadPoolExecutor(maximumParallelSends
+                + MAX_SKELETON_THREADS + MQTT_THREADS, schedulerNamedThreadFactory);
         scheduler.setKeepAliveTime(100, TimeUnit.SECONDS);
         scheduler.allowCoreThreadTimeOut(true);
         return scheduler;