+++ java/messaging/messaging-common/src/main/java/io/joynr/messaging/ConfigurableMessagingSettings.java
@@ -62,6 +62,7 @@ public class ConfigurableMessagingSettings implements MessagingSettings {
     public static final String PROPERTY_HOSTS_FILENAME = "joynr.messaging.hostsfilename";
 
     public static final String PROPERTY_MAX_MESSAGE_SIZE = "joynr.messaging.maxmessagesize";
+    public static final String PROPERTY_MAX_MESSAGES_INQUEUE = "joynr.messaging.maxmessagesinqueue";
 
     public static final String PROPERTY_MESSAGING_MAXIMUM_TTL_MS = "joynr.messaging.maxttlms";
     public static final String PROPERTY_TTL_UPLIFT_MS = "joynr.messaging.ttlupliftms";